API-AI-vehicle- project/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.3
exceptiongroup==1.3.1
fastapi==0.124.0
h11==0.16.0
idna==3.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
websockets==15.0.1
python-dotenv==1.0.0
